refactor: make the missing plugin error more generic
So it can be used also for custom 3rd-party plugins.
